requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
discord.py==1.0.1
dnspython==1.16.0
idna==2.8
multidict==4.5.2
Pillow==6.0.0
psutil==5.6.1
pycparser==2.19
pymongo[tls,srv]==3.7.2
PyNaCl==1.3.0
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
websockets==6.0
lavalink==2.1.10
yarl==1.3.0
